backend_app/requirements.txt
flask
flask-cors
flask-sqlalchemy
python-dotenv
qrcode
reportlab
web3
gunicorn
